refactor: move decryption code to `ConversationGuard` [WPB-15583]
Also, take the opportunity to make the naming
consistent in the `buffer_commit` module:
pending commit -> buffered commit
pending buffered commit -> buffered commit
This is a better name because of name collision with
pending commits in `openmls`.
