+++ crypto/src/mls/conversation/conversation_guard/mod.rs
@@ -1,5 +1,5 @@
 mod commit;
-mod decrypt;
+pub(crate) mod decrypt;
 mod encrypt;
 mod merge;
